Fix: Allow anonymous download when NoAnonymousUploads is set
When the option NoAnonymousUploads is set, any access to the web front end page will result in a login page
even if the login is not requested by the server.
This is an issue when a user wants to assess a download page (in /#?id=dowloadId)
So this commit disables the login check in the plik frontend client if the page is a download page.