feat: add compatibility with OpenShift
OpenShift is setting a random user at statup of a container. The only thing known is that this user is member of the `root` group. This commit is about adding the `plik` user to the `root` group, and then set all the files to be owned by the `root` group.

This makes the image compatible with OpenShift.